requirements.txt
jupyter
pytest
matplotlib>=3.2.1
pandas>=1.0.0
scikit-learn==0.23.1
scipy==1.4.1
seaborn>=0.10.0
umap-learn>=0.3.10
shap
jupyter
jupyter_contrib_nbextensions
